Add: ARIES native build support and cleanup for cpp tutorials
Make runtime/cpp/image_classification and runtime/cpp/object_detection
build and run on both ARIES (x86_64 native, NPU on host) and REGULUS
(aarch64 cross-compile, NPU on target board) from the same
CMakeLists.txt. Each CMakeLists now branches on CMAKE_SYSTEM_PROCESSOR
to pick the right `-march` flag, and the model load path passes an
explicit single-core ModelConfig so ARIES multi-mode mxq
(inference_scheme="all" from model_compile.py) loads alongside REGULUS
single-mode mxq. README.{md,KR.md} document the apt OpenCV prereq for
the ARIES path and the two distinct run commands.

In object_detection, also: drop utils/inference/parser.cc and remove
parse_yaml() from the header since the tutorial builds ModelInfo from
hardcoded `make_yolo_detect_config()` and never calls the YAML parser;
rename `parser.h` -> `types.h` so the header reflects its actual role
as a shared type definitions header (PreProcessOps, Task, ImageSize,
PreProcessInfo, PostProcessInfo, ModelInfo); rename `yolov9m_config.h`
-> `yolo_detect_config.h` to reflect that the same config covers both
YOLO11m (ARIES) and YOLOv9m (REGULUS) (ultralytics v8/v9/v11 anchor-free
Detect heads share an output layout of 3 strides x [reg_max*4 box +
nc cls]); and rewrite all file-level and non-obvious inline comments
into bilingual form (English first, then a Korean paragraph or
parenthetical).

(KR) cpp 튜토리얼의 ARIES 네이티브 빌드 지원과 코드 정리.

runtime/cpp/image_classification 와 runtime/cpp/object_detection 가
같은 CMakeLists.txt 로 ARIES (x86_64 네이티브, 호스트 NPU) 와 REGULUS
(aarch64 크로스 컴파일, 보드 NPU) 양쪽에서 빌드·실행되도록 정리.
각 CMakeLists 가 CMAKE_SYSTEM_PROCESSOR 를 분기해 `-march` 를
자동 선택하고, 모델 로드 시 ModelConfig 의 single-core 모드를 명시해
ARIES 의 multi-mode mxq (model_compile.py 의 inference_scheme="all")
와 REGULUS 의 single-mode mxq 가 같은 코드로 로드되도록 함.
README.{md,KR.md} 가 ARIES 경로의 apt OpenCV 사전 설치 명령과 두
실행 명령을 명시.

object_detection 에서는 추가로: 하드코드 `make_yolo_detect_config()`
가 ModelInfo 를 만들어 YAML 파서가 한 번도 호출되지 않으므로
utils/inference/parser.cc 를 삭제하고 헤더에서 parse_yaml() 선언을
제거. `parser.h` 가 실제로는 자료형 헤더 (PreProcessOps, Task,
ImageSize, PreProcessInfo, PostProcessInfo, ModelInfo) 라는 역할에
맞춰 `types.h` 로 rename. `yolov9m_config.h` 는 같은 설정이 YOLO11m
(ARIES) 와 YOLOv9m (REGULUS) 둘 다 적용된다는 사실을 반영해
`yolo_detect_config.h` 로 rename (ultralytics v8/v9/v11 의
anchor-free Detect head 는 3 stride x [reg_max*4 box + nc cls] 출력
layout 을 공유). 모든 파일 헤더와 비명확 inline 주석을 영어/한국어
이중 표기로 통일.
