+++ runtime/cpp/image_classification/README.md
@@ -1,80 +1,87 @@
-# Image Classification - C++ Cross-Compiled Inference
+# Image Classification - C++ Inference (ARIES + REGULUS)
 
 An example of running C++ NPU inference on a single image using a ResNet-50 MXQ model.
+Supports both **ARIES native build** (x86_64 host with NPU) and **REGULUS cross-compile**
+(x86_64 host -> ARM64 target board) from the same `CMakeLists.txt`.
 
 ## File Structure
 
 - `infer_cls.cc` - Inference binary source (NPU inference, Top-5 output)
-- `CMakeLists.txt` - CMake build configuration
+- `CMakeLists.txt` - CMake build configuration (host arch auto-detected)
 - `imagenet_labels.txt` - ImageNet 1000 class label file
 
 ## Prerequisites
 
-- You need a `resnet50.mxq` file generated with `model_compile_regulus.py` from the REGULUS section of the [compiler tutorial](../../../compilation/image_classification/README.md).
+- `resnet50.mxq` file generated with the matching compiler tutorial
+  ([../../../compilation/image_classification/README.md](../../../compilation/image_classification/README.md)).
+  ARIES uses `model_compile.py`, REGULUS uses `model_compile_regulus.py`.
 
-### 1. Verify Toolchain Installation
+### Common requirements (both paths)
 
-Check that the cross-compilation toolchain is installed:
+- CMake >= 3.18
+- C++17 compiler (gcc / clang)
+- `qbruntime` library (installed together with the Mobilint NPU SDK)
 
-```bash
-ls /opt/crosstools/mobilint/
-```
+### ARIES native build (x86_64 host with NPU)
 
-You should see a version directory (e.g., `1.0.0/`). If the toolchain is not installed, refer to [Cross-Compilation Setup](../README.md) to install it first.
-
-### 2. Activate Cross-Compilation Environment
-
-Find the `environment-setup-cortexa53-mobilint-linux` file in the installed toolchain directory:
+Install host-side OpenCV and build tools (Ubuntu / Debian):
 
 ```bash
-find /opt/crosstools/mobilint/ -name "environment-setup-cortexa53-mobilint-linux"
-# /opt/crosstools/mobilint/1.0.0/v3.4.0/environment-setup-cortexa53-mobilint-linux
+apt-get update
+apt-get install -y build-essential cmake libopencv-dev
 ```
 
-Unset host library paths (e.g., CUDA) to prevent conflicts with the cross-compiler:
+### REGULUS cross-compile (x86_64 host -> ARM64 target board)
+
+The vendor cross-compile toolchain ships with OpenCV and `qbruntime` pre-installed.
+Verify the toolchain and activate it:
 
 ```bash
-unset LD_LIBRARY_PATH
+ls /opt/crosstools/mobilint/                                   # version directory expected
+unset LD_LIBRARY_PATH                                          # avoid host CUDA libs leaking
+source /opt/crosstools/mobilint/{version}/{sdk}/environment-setup-cortexa53-mobilint-linux
+echo $CXX                                                      # aarch64-mobilint-linux-g++ ...
 ```
 
-Activate the toolchain environment using the path found above:
+If the toolchain is not installed, follow [Cross-Compilation Setup](../README.md).
+
+## Build
+
+The same command works for both ARIES native and REGULUS cross-compile. The
+`CMakeLists.txt` detects the host arch and selects the right `-march` flag.
 
 ```bash
-source /opt/crosstools/mobilint/1.0.0/v3.4.0/environment-setup-cortexa53-mobilint-linux
+cmake -B build -S .
+cmake --build build -j
 ```
 
-Verify that the cross-compiler is set:
+After a successful build, `build/infer-cls` is created.
 
-```bash
-echo $CXX
-# aarch64-mobilint-linux-g++ ...
-```
-
-## Build (Host)
-
-```bash
-mkdir build && cd build
-cmake ..
-make -j8
-cd ..
-```
-
-After a successful build, the `build/infer-cls` binary will be generated.
-
-Verify the binary is built for ARM64:
+Verify the architecture:
 
 ```bash
 file build/infer-cls
-# build/infer-cls: ELF 64-bit LSB executable, ARM aarch64, ...
+# ARIES:   ELF 64-bit LSB executable, x86-64, ...
+# REGULUS: ELF 64-bit LSB executable, ARM aarch64, ...
 ```
 
-## Run (Target Board)
+## Run
 
-Copy the `build/infer-cls` binary, `resnet50.mxq` model file, and `imagenet_labels.txt` label file to the target board, then run:
+A sample image `../rc/volcano.jpg` is bundled with the repo.
+
+### ARIES (same host)
+
+```bash
+./build/infer-cls ../../../compilation/image_classification/resnet50.mxq ../rc/volcano.jpg imagenet_labels.txt
+```
+
+### REGULUS (target board)
+
+Copy `build/infer-cls`, `resnet50.mxq`, `imagenet_labels.txt`, and `../rc/volcano.jpg` to the target board, then:
 
 ```bash
 chmod +x infer-cls
-./infer-cls resnet50.mxq example.jpg imagenet_labels.txt
+./infer-cls resnet50.mxq volcano.jpg imagenet_labels.txt
 ```
 
 ## Example Output
