+++ runtime/cpp/object_detection/utils/inference/runner.h
@@ -1,10 +1,8 @@
-// =============================================================================
-// runner.h - NPU 실행 래퍼 (NPU runner) + .npy 작성 유틸
-// =============================================================================
-// qbruntime Accelerator/Model 을 감싸는 클래스. 타겟 보드(ARM64)에서만 NPU 접근.
+// NPU runner wrapping qbruntime Accelerator and Model, plus a NumPy .npy save utility.
+// NPU access is only available on the target ARM64 board; host builds will link-fail on qbruntime symbols.
 //
-// Wraps qbruntime Accelerator/Model. NPU access only on the target board (ARM64).
-// =============================================================================
+// (KR) qbruntime Accelerator 와 Model 을 래핑하는 NPU 실행기, NumPy .npy 저장 유틸 포함.
+// NPU 접근은 대상 ARM64 보드에서만 가능하며, 호스트 빌드는 qbruntime 심벌에서 링크 오류가 발생한다.
 #pragma once
 #include <qbruntime/qbruntime.h>
 
@@ -18,13 +16,16 @@ public:
     explicit NPURunner(const std::string& model_path);
     ~NPURunner();
 
-    // float 입력 추론. HWC float 배열을 받아 출력 텐서 벡터 반환.
+    // Runs inference with a float HWC input buffer; returns output tensors as flat float vectors.
+    // (KR: float HWC 입력 버퍼로 추론 실행; 출력 텐서를 평면 float 벡터 목록으로 반환.)
     std::vector<std::vector<float>> infer(std::unique_ptr<float[]> input);
 
-    // uint8 입력 추론 (uint8-input compiled MXQ).
+    // Runs inference with a uint8 CHW input buffer compiled with Uint8InputConfig.
+    // (KR: Uint8InputConfig 로 컴파일된 MXQ 에 대해 uint8 CHW 입력 버퍼로 추론 실행.)
     std::vector<std::vector<float>> infer_uint8(std::unique_ptr<uint8_t[]> input);
 
-    // 모델 입력 shape: {height, width, channel}
+    // Returns model input shape as {height, width, channel}.
+    // (KR: 모델 입력 shape 를 {height, width, channel} 순서로 반환.)
     std::vector<int> get_input_shape() const;
 
 private:
@@ -33,10 +34,10 @@ private:
     std::unique_ptr<mobilint::Model> model_;
 };
 
-// .npy (numpy 표준) 단일 텐서 저장. zip / 외부 라이브러리 없음.
-// 포맷: magic(\x93NUMPY) + ver(1.0) + JSON 헤더 + binary float32.
-// path: 출력 파일 경로
-// data: 평면 float32 배열 (row-major)
-// shape: 차원 순서 그대로 (예: {1, 64, 80, 80})
+// Saves a single float32 tensor to a .npy file (NumPy v1.0 format, no zip, no external libraries).
+// Format: magic(\x93NUMPY) + version(1.0) + ASCII dict header + raw float32 data row-major.
+// shape entries follow NumPy dimension order (e.g., {1, 64, 80, 80}).
+// (KR: 단일 float32 텐서를 .npy 파일로 저장(NumPy v1.0, zip 없음, 외부 라이브러리 없음).
+// shape 는 NumPy 차원 순서를 따른다(예: {1, 64, 80, 80}).)
 void save_npy(const std::string& path, const float* data,
               const std::vector<size_t>& shape);
